Find All Markers at a range of resolutions

Find All Markers at a range of resolutions

Usage

find_all_markers(seu, metavar = NULL, seurat_assay = "gene", ...)

Arguments

metavar

Examples

markers_stashed_seu <- find_all_markers(panc8)
#> stashing presto markers for gene_snn_res.0.2
#> stashing genesorteR markers for gene_snn_res.0.2
#> Warning: A Friendly Warning: Some genes were removed because they were zeros in all cells after binarization. You probably don't need to do anything but you might want to look into this. Maybe you forgot to pre-filter the genes? You can also use a different binarization method. Excluded genes are available in the output under '$removed'.
#> stashing presto markers for gene_snn_res.0.4
#> stashing genesorteR markers for gene_snn_res.0.4
#> Warning: A Friendly Warning: Some genes were removed because they were zeros in all cells after binarization. You probably don't need to do anything but you might want to look into this. Maybe you forgot to pre-filter the genes? You can also use a different binarization method. Excluded genes are available in the output under '$removed'.
#> stashing presto markers for gene_snn_res.0.6
#> stashing genesorteR markers for gene_snn_res.0.6
#> Warning: A Friendly Warning: Some genes were removed because they were zeros in all cells after binarization. You probably don't need to do anything but you might want to look into this. Maybe you forgot to pre-filter the genes? You can also use a different binarization method. Excluded genes are available in the output under '$removed'.
#> stashing presto markers for gene_snn_res.0.8
#> stashing genesorteR markers for gene_snn_res.0.8
#> Warning: A Friendly Warning: Some genes were removed because they were zeros in all cells after binarization. You probably don't need to do anything but you might want to look into this. Maybe you forgot to pre-filter the genes? You can also use a different binarization method. Excluded genes are available in the output under '$removed'.
#> stashing presto markers for gene_snn_res.1
#> stashing genesorteR markers for gene_snn_res.1
#> Warning: A Friendly Warning: Some genes were removed because they were zeros in all cells after binarization. You probably don't need to do anything but you might want to look into this. Maybe you forgot to pre-filter the genes? You can also use a different binarization method. Excluded genes are available in the output under '$removed'.
#> stashing presto markers for gene_snn_res.1.2
#> stashing genesorteR markers for gene_snn_res.1.2
#> Warning: A Friendly Warning: Some genes were removed because they were zeros in all cells after binarization. You probably don't need to do anything but you might want to look into this. Maybe you forgot to pre-filter the genes? You can also use a different binarization method. Excluded genes are available in the output under '$removed'.
#> stashing presto markers for gene_snn_res.1.4
#> stashing genesorteR markers for gene_snn_res.1.4
#> Warning: A Friendly Warning: Some genes were removed because they were zeros in all cells after binarization. You probably don't need to do anything but you might want to look into this. Maybe you forgot to pre-filter the genes? You can also use a different binarization method. Excluded genes are available in the output under '$removed'.
#> stashing presto markers for gene_snn_res.1.6
#> stashing genesorteR markers for gene_snn_res.1.6
#> Warning: A Friendly Warning: Some genes were removed because they were zeros in all cells after binarization. You probably don't need to do anything but you might want to look into this. Maybe you forgot to pre-filter the genes? You can also use a different binarization method. Excluded genes are available in the output under '$removed'.
#> stashing presto markers for gene_snn_res.1.8
#> stashing genesorteR markers for gene_snn_res.1.8
#> Warning: A Friendly Warning: Some genes were removed because they were zeros in all cells after binarization. You probably don't need to do anything but you might want to look into this. Maybe you forgot to pre-filter the genes? You can also use a different binarization method. Excluded genes are available in the output under '$removed'.
#> stashing presto markers for gene_snn_res.2
#> stashing genesorteR markers for gene_snn_res.2
#> Warning: A Friendly Warning: Some genes were removed because they were zeros in all cells after binarization. You probably don't need to do anything but you might want to look into this. Maybe you forgot to pre-filter the genes? You can also use a different binarization method. Excluded genes are available in the output under '$removed'.
marker_genes <- Misc(markers_stashed_seu, "markers")
str(marker_genes)
#> List of 10
#>  $ gene_snn_res.0.2:List of 2
#>   ..$ presto    : gropd_df [1,200 × 5] (S3: grouped_df/tbl_df/tbl/data.frame)
#>   .. ..$ Gene.Name              : chr [1:1200] "GCG" "TTR" "CHGB" "TM4SF4" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:1200] 5.12 3.84 2.55 2.42 2.18 ...
#>   .. ..$ Adjusted.pvalue        : num [1:1200] 2.53e-88 9.32e-124 4.93e-89 3.22e-94 1.37e-88 ...
#>   .. ..$ avgExpr                : num [1:1200] 6.92 6.44 4.71 3.57 3.25 ...
#>   .. ..$ Cluster                : chr [1:1200] "0" "0" "0" "0" ...
#>   .. ..- attr(*, "groups")= tibble [6 × 2] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ Cluster: chr [1:6] "0" "1" "2" "3" ...
#>   .. .. ..$ .rows  : list<int> [1:6] 
#>   .. .. .. ..$ : int [1:200] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. .. ..$ : int [1:200] 201 202 203 204 205 206 207 208 209 210 ...
#>   .. .. .. ..$ : int [1:200] 401 402 403 404 405 406 407 408 409 410 ...
#>   .. .. .. ..$ : int [1:200] 601 602 603 604 605 606 607 608 609 610 ...
#>   .. .. .. ..$ : int [1:200] 801 802 803 804 805 806 807 808 809 810 ...
#>   .. .. .. ..$ : int [1:200] 1001 1002 1003 1004 1005 1006 1007 1008 1009 1010 ...
#>   .. .. .. ..@ ptype: int(0) 
#>   .. .. ..- attr(*, ".drop")= logi TRUE
#>   ..$ genesorteR:'data.frame':	14850 obs. of  5 variables:
#>   .. ..$ Gene.Name              : chr [1:14850] "ALDH1A3" "VSTM4" "LRRC32" "COL18A1" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:14850] 0.09603 0.0036 0.00254 1.77954 0.0017 ...
#>   .. ..$ Adjusted.pvalue        : num [1:14850] 1.24e-05 1.24e-05 1.24e-05 1.24e-05 1.24e-05 ...
#>   .. ..$ Specificity.Score      : num [1:14850] 0.708 0.698 0.642 0.625 0.58 ...
#>   .. ..$ Cluster                : chr [1:14850] "3" "5" "5" "3" ...
#>  $ gene_snn_res.0.4:List of 2
#>   ..$ presto    : gropd_df [1,600 × 5] (S3: grouped_df/tbl_df/tbl/data.frame)
#>   .. ..$ Gene.Name              : chr [1:1600] "GCG" "TTR" "CHGB" "TM4SF4" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:1600] 5.11 3.84 2.54 2.44 2.19 ...
#>   .. ..$ Adjusted.pvalue        : num [1:1600] 8.94e-87 1.90e-122 4.38e-87 2.08e-95 1.71e-87 ...
#>   .. ..$ avgExpr                : num [1:1600] 6.97 6.49 4.74 3.61 3.28 ...
#>   .. ..$ Cluster                : chr [1:1600] "0" "0" "0" "0" ...
#>   .. ..- attr(*, "groups")= tibble [8 × 2] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ Cluster: chr [1:8] "0" "1" "2" "3" ...
#>   .. .. ..$ .rows  : list<int> [1:8] 
#>   .. .. .. ..$ : int [1:200] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. .. ..$ : int [1:200] 201 202 203 204 205 206 207 208 209 210 ...
#>   .. .. .. ..$ : int [1:200] 401 402 403 404 405 406 407 408 409 410 ...
#>   .. .. .. ..$ : int [1:200] 601 602 603 604 605 606 607 608 609 610 ...
#>   .. .. .. ..$ : int [1:200] 801 802 803 804 805 806 807 808 809 810 ...
#>   .. .. .. ..$ : int [1:200] 1001 1002 1003 1004 1005 1006 1007 1008 1009 1010 ...
#>   .. .. .. ..$ : int [1:200] 1201 1202 1203 1204 1205 1206 1207 1208 1209 1210 ...
#>   .. .. .. ..$ : int [1:200] 1401 1402 1403 1404 1405 1406 1407 1408 1409 1410 ...
#>   .. .. .. ..@ ptype: int(0) 
#>   .. .. ..- attr(*, ".drop")= logi TRUE
#>   ..$ genesorteR:'data.frame':	17847 obs. of  5 variables:
#>   .. ..$ Gene.Name              : chr [1:17847] "DEFB1" "TRPV6" "PROM1" "SERPINA5" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:17847] 0.0271 0.0203 0.0161 0.1832 0.2203 ...
#>   .. ..$ Adjusted.pvalue        : num [1:17847] 1.26e-05 1.26e-05 1.26e-05 1.26e-05 1.26e-05 ...
#>   .. ..$ Specificity.Score      : num [1:17847] 0.711 0.71 0.639 0.632 0.627 ...
#>   .. ..$ Cluster                : chr [1:17847] "1" "1" "1" "1" ...
#>  $ gene_snn_res.0.6:List of 2
#>   ..$ presto    : gropd_df [1,800 × 5] (S3: grouped_df/tbl_df/tbl/data.frame)
#>   .. ..$ Gene.Name              : chr [1:1800] "GCG" "TTR" "CHGB" "TM4SF4" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:1800] 5.11 3.84 2.54 2.44 2.19 ...
#>   .. ..$ Adjusted.pvalue        : num [1:1800] 8.94e-87 1.90e-122 4.38e-87 2.08e-95 1.71e-87 ...
#>   .. ..$ avgExpr                : num [1:1800] 6.97 6.49 4.74 3.61 3.28 ...
#>   .. ..$ Cluster                : chr [1:1800] "0" "0" "0" "0" ...
#>   .. ..- attr(*, "groups")= tibble [9 × 2] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ Cluster: chr [1:9] "0" "1" "2" "3" ...
#>   .. .. ..$ .rows  : list<int> [1:9] 
#>   .. .. .. ..$ : int [1:200] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. .. ..$ : int [1:200] 201 202 203 204 205 206 207 208 209 210 ...
#>   .. .. .. ..$ : int [1:200] 401 402 403 404 405 406 407 408 409 410 ...
#>   .. .. .. ..$ : int [1:200] 601 602 603 604 605 606 607 608 609 610 ...
#>   .. .. .. ..$ : int [1:200] 801 802 803 804 805 806 807 808 809 810 ...
#>   .. .. .. ..$ : int [1:200] 1001 1002 1003 1004 1005 1006 1007 1008 1009 1010 ...
#>   .. .. .. ..$ : int [1:200] 1201 1202 1203 1204 1205 1206 1207 1208 1209 1210 ...
#>   .. .. .. ..$ : int [1:200] 1401 1402 1403 1404 1405 1406 1407 1408 1409 1410 ...
#>   .. .. .. ..$ : int [1:200] 1601 1602 1603 1604 1605 1606 1607 1608 1609 1610 ...
#>   .. .. .. ..@ ptype: int(0) 
#>   .. .. ..- attr(*, ".drop")= logi TRUE
#>   ..$ genesorteR:'data.frame':	18326 obs. of  5 variables:
#>   .. ..$ Gene.Name              : chr [1:18326] "PPAP2C" "DEFB1" "TRPV6" "LGALS4" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:18326] 0.03161 0.07263 0.05734 0.03596 0.00895 ...
#>   .. ..$ Adjusted.pvalue        : num [1:18326] 1.31e-05 1.31e-05 1.31e-05 1.31e-05 1.31e-05 ...
#>   .. ..$ Specificity.Score      : num [1:18326] 0.75 0.725 0.724 0.692 0.667 ...
#>   .. ..$ Cluster                : chr [1:18326] "1" "1" "1" "1" ...
#>  $ gene_snn_res.0.8:List of 2
#>   ..$ presto    : gropd_df [1,800 × 5] (S3: grouped_df/tbl_df/tbl/data.frame)
#>   .. ..$ Gene.Name              : chr [1:1800] "GCG" "TTR" "CHGB" "TM4SF4" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:1800] 5.01 3.67 2.41 2.31 2.18 ...
#>   .. ..$ Adjusted.pvalue        : num [1:1800] 1.99e-77 9.72e-105 5.66e-72 8.42e-79 4.88e-79 ...
#>   .. ..$ avgExpr                : num [1:1800] 7.11 6.53 4.75 3.61 3.36 ...
#>   .. ..$ Cluster                : chr [1:1800] "0" "0" "0" "0" ...
#>   .. ..- attr(*, "groups")= tibble [9 × 2] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ Cluster: chr [1:9] "0" "1" "2" "3" ...
#>   .. .. ..$ .rows  : list<int> [1:9] 
#>   .. .. .. ..$ : int [1:200] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. .. ..$ : int [1:200] 201 202 203 204 205 206 207 208 209 210 ...
#>   .. .. .. ..$ : int [1:200] 401 402 403 404 405 406 407 408 409 410 ...
#>   .. .. .. ..$ : int [1:200] 601 602 603 604 605 606 607 608 609 610 ...
#>   .. .. .. ..$ : int [1:200] 801 802 803 804 805 806 807 808 809 810 ...
#>   .. .. .. ..$ : int [1:200] 1001 1002 1003 1004 1005 1006 1007 1008 1009 1010 ...
#>   .. .. .. ..$ : int [1:200] 1201 1202 1203 1204 1205 1206 1207 1208 1209 1210 ...
#>   .. .. .. ..$ : int [1:200] 1401 1402 1403 1404 1405 1406 1407 1408 1409 1410 ...
#>   .. .. .. ..$ : int [1:200] 1601 1602 1603 1604 1605 1606 1607 1608 1609 1610 ...
#>   .. .. .. ..@ ptype: int(0) 
#>   .. .. ..- attr(*, ".drop")= logi TRUE
#>   ..$ genesorteR:'data.frame':	16728 obs. of  5 variables:
#>   .. ..$ Gene.Name              : chr [1:16728] "PPAP2C" "DEFB1" "TRPV6" "LGALS4" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:16728] 0.03161 0.07263 0.05734 0.03596 0.00895 ...
#>   .. ..$ Adjusted.pvalue        : num [1:16728] 1.3e-05 1.3e-05 1.3e-05 1.3e-05 1.3e-05 ...
#>   .. ..$ Specificity.Score      : num [1:16728] 0.75 0.725 0.724 0.692 0.667 ...
#>   .. ..$ Cluster                : chr [1:16728] "1" "1" "1" "1" ...
#>  $ gene_snn_res.1  :List of 2
#>   ..$ presto    : gropd_df [1,800 × 5] (S3: grouped_df/tbl_df/tbl/data.frame)
#>   .. ..$ Gene.Name              : chr [1:1800] "GCG" "TTR" "CHGB" "TM4SF4" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:1800] 5.01 3.67 2.41 2.31 2.18 ...
#>   .. ..$ Adjusted.pvalue        : num [1:1800] 1.99e-77 9.72e-105 5.66e-72 8.42e-79 4.88e-79 ...
#>   .. ..$ avgExpr                : num [1:1800] 7.11 6.53 4.75 3.61 3.36 ...
#>   .. ..$ Cluster                : chr [1:1800] "0" "0" "0" "0" ...
#>   .. ..- attr(*, "groups")= tibble [9 × 2] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ Cluster: chr [1:9] "0" "1" "2" "3" ...
#>   .. .. ..$ .rows  : list<int> [1:9] 
#>   .. .. .. ..$ : int [1:200] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. .. ..$ : int [1:200] 201 202 203 204 205 206 207 208 209 210 ...
#>   .. .. .. ..$ : int [1:200] 401 402 403 404 405 406 407 408 409 410 ...
#>   .. .. .. ..$ : int [1:200] 601 602 603 604 605 606 607 608 609 610 ...
#>   .. .. .. ..$ : int [1:200] 801 802 803 804 805 806 807 808 809 810 ...
#>   .. .. .. ..$ : int [1:200] 1001 1002 1003 1004 1005 1006 1007 1008 1009 1010 ...
#>   .. .. .. ..$ : int [1:200] 1201 1202 1203 1204 1205 1206 1207 1208 1209 1210 ...
#>   .. .. .. ..$ : int [1:200] 1401 1402 1403 1404 1405 1406 1407 1408 1409 1410 ...
#>   .. .. .. ..$ : int [1:200] 1601 1602 1603 1604 1605 1606 1607 1608 1609 1610 ...
#>   .. .. .. ..@ ptype: int(0) 
#>   .. .. ..- attr(*, ".drop")= logi TRUE
#>   ..$ genesorteR:'data.frame':	16728 obs. of  5 variables:
#>   .. ..$ Gene.Name              : chr [1:16728] "PPAP2C" "DEFB1" "TRPV6" "LGALS4" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:16728] 0.03161 0.07263 0.05734 0.03596 0.00895 ...
#>   .. ..$ Adjusted.pvalue        : num [1:16728] 1.3e-05 1.3e-05 1.3e-05 1.3e-05 1.3e-05 ...
#>   .. ..$ Specificity.Score      : num [1:16728] 0.75 0.725 0.724 0.692 0.667 ...
#>   .. ..$ Cluster                : chr [1:16728] "1" "1" "1" "1" ...
#>  $ gene_snn_res.1.2:List of 2
#>   ..$ presto    : gropd_df [2,200 × 5] (S3: grouped_df/tbl_df/tbl/data.frame)
#>   .. ..$ Gene.Name              : chr [1:2200] "GCG" "TTR" "CHGB" "CRYBA2" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:2200] 4.88 3.45 2.39 2.29 2.28 ...
#>   .. ..$ Adjusted.pvalue        : num [1:2200] 8.47e-63 9.08e-80 1.13e-61 4.40e-87 1.33e-63 ...
#>   .. ..$ avgExpr                : num [1:2200] 7.34 6.59 4.89 2.87 3.74 ...
#>   .. ..$ Cluster                : chr [1:2200] "0" "0" "0" "0" ...
#>   .. ..- attr(*, "groups")= tibble [11 × 2] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ Cluster: chr [1:11] "0" "1" "10" "2" ...
#>   .. .. ..$ .rows  : list<int> [1:11] 
#>   .. .. .. ..$ : int [1:200] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. .. ..$ : int [1:200] 201 202 203 204 205 206 207 208 209 210 ...
#>   .. .. .. ..$ : int [1:200] 401 402 403 404 405 406 407 408 409 410 ...
#>   .. .. .. ..$ : int [1:200] 601 602 603 604 605 606 607 608 609 610 ...
#>   .. .. .. ..$ : int [1:200] 801 802 803 804 805 806 807 808 809 810 ...
#>   .. .. .. ..$ : int [1:200] 1001 1002 1003 1004 1005 1006 1007 1008 1009 1010 ...
#>   .. .. .. ..$ : int [1:200] 1201 1202 1203 1204 1205 1206 1207 1208 1209 1210 ...
#>   .. .. .. ..$ : int [1:200] 1401 1402 1403 1404 1405 1406 1407 1408 1409 1410 ...
#>   .. .. .. ..$ : int [1:200] 1601 1602 1603 1604 1605 1606 1607 1608 1609 1610 ...
#>   .. .. .. ..$ : int [1:200] 1801 1802 1803 1804 1805 1806 1807 1808 1809 1810 ...
#>   .. .. .. ..$ : int [1:200] 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 ...
#>   .. .. .. ..@ ptype: int(0) 
#>   .. .. ..- attr(*, ".drop")= logi TRUE
#>   ..$ genesorteR:'data.frame':	21978 obs. of  5 variables:
#>   .. ..$ Gene.Name              : chr [1:21978] "PPAP2C" "DEFB1" "TRPV6" "LGALS4" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:21978] 0.03161 0.07263 0.05734 0.03596 0.00895 ...
#>   .. ..$ Adjusted.pvalue        : num [1:21978] 1.33e-05 1.33e-05 1.33e-05 1.33e-05 1.33e-05 ...
#>   .. ..$ Specificity.Score      : num [1:21978] 0.75 0.725 0.724 0.692 0.667 ...
#>   .. ..$ Cluster                : chr [1:21978] "1" "1" "1" "1" ...
#>  $ gene_snn_res.1.4:List of 2
#>   ..$ presto    : gropd_df [2,400 × 5] (S3: grouped_df/tbl_df/tbl/data.frame)
#>   .. ..$ Gene.Name              : chr [1:2400] "GCG" "TTR" "CHGB" "TM4SF4" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:2400] 4.65 3.35 2.31 2.25 2.23 ...
#>   .. ..$ Adjusted.pvalue        : num [1:2400] 9.24e-49 3.18e-67 3.60e-50 2.66e-54 1.62e-70 ...
#>   .. ..$ avgExpr                : num [1:2400] 7.32 6.63 4.91 3.79 2.9 ...
#>   .. ..$ Cluster                : chr [1:2400] "0" "0" "0" "0" ...
#>   .. ..- attr(*, "groups")= tibble [12 × 2] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ Cluster: chr [1:12] "0" "1" "10" "11" ...
#>   .. .. ..$ .rows  : list<int> [1:12] 
#>   .. .. .. ..$ : int [1:200] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. .. ..$ : int [1:200] 201 202 203 204 205 206 207 208 209 210 ...
#>   .. .. .. ..$ : int [1:200] 401 402 403 404 405 406 407 408 409 410 ...
#>   .. .. .. ..$ : int [1:200] 601 602 603 604 605 606 607 608 609 610 ...
#>   .. .. .. ..$ : int [1:200] 801 802 803 804 805 806 807 808 809 810 ...
#>   .. .. .. ..$ : int [1:200] 1001 1002 1003 1004 1005 1006 1007 1008 1009 1010 ...
#>   .. .. .. ..$ : int [1:200] 1201 1202 1203 1204 1205 1206 1207 1208 1209 1210 ...
#>   .. .. .. ..$ : int [1:200] 1401 1402 1403 1404 1405 1406 1407 1408 1409 1410 ...
#>   .. .. .. ..$ : int [1:200] 1601 1602 1603 1604 1605 1606 1607 1608 1609 1610 ...
#>   .. .. .. ..$ : int [1:200] 1801 1802 1803 1804 1805 1806 1807 1808 1809 1810 ...
#>   .. .. .. ..$ : int [1:200] 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 ...
#>   .. .. .. ..$ : int [1:200] 2201 2202 2203 2204 2205 2206 2207 2208 2209 2210 ...
#>   .. .. .. ..@ ptype: int(0) 
#>   .. .. ..- attr(*, ".drop")= logi TRUE
#>   ..$ genesorteR:'data.frame':	24734 obs. of  5 variables:
#>   .. ..$ Gene.Name              : chr [1:24734] "PCAT19" "PPAP2C" "DEFB1" "TRPV6" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:24734] 0.00287 0.03161 0.07263 0.05734 0.00077 ...
#>   .. ..$ Adjusted.pvalue        : num [1:24734] 1.34e-05 1.34e-05 1.34e-05 1.34e-05 1.34e-05 ...
#>   .. ..$ Specificity.Score      : num [1:24734] 0.805 0.755 0.73 0.728 0.71 ...
#>   .. ..$ Cluster                : chr [1:24734] "11" "1" "1" "1" ...
#>  $ gene_snn_res.1.6:List of 2
#>   ..$ presto    : gropd_df [2,400 × 5] (S3: grouped_df/tbl_df/tbl/data.frame)
#>   .. ..$ Gene.Name              : chr [1:2400] "GCG" "TTR" "CHGB" "TM4SF4" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:2400] 4.65 3.35 2.31 2.25 2.23 ...
#>   .. ..$ Adjusted.pvalue        : num [1:2400] 9.24e-49 3.18e-67 3.60e-50 2.66e-54 1.62e-70 ...
#>   .. ..$ avgExpr                : num [1:2400] 7.32 6.63 4.91 3.79 2.9 ...
#>   .. ..$ Cluster                : chr [1:2400] "0" "0" "0" "0" ...
#>   .. ..- attr(*, "groups")= tibble [12 × 2] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ Cluster: chr [1:12] "0" "1" "10" "11" ...
#>   .. .. ..$ .rows  : list<int> [1:12] 
#>   .. .. .. ..$ : int [1:200] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. .. ..$ : int [1:200] 201 202 203 204 205 206 207 208 209 210 ...
#>   .. .. .. ..$ : int [1:200] 401 402 403 404 405 406 407 408 409 410 ...
#>   .. .. .. ..$ : int [1:200] 601 602 603 604 605 606 607 608 609 610 ...
#>   .. .. .. ..$ : int [1:200] 801 802 803 804 805 806 807 808 809 810 ...
#>   .. .. .. ..$ : int [1:200] 1001 1002 1003 1004 1005 1006 1007 1008 1009 1010 ...
#>   .. .. .. ..$ : int [1:200] 1201 1202 1203 1204 1205 1206 1207 1208 1209 1210 ...
#>   .. .. .. ..$ : int [1:200] 1401 1402 1403 1404 1405 1406 1407 1408 1409 1410 ...
#>   .. .. .. ..$ : int [1:200] 1601 1602 1603 1604 1605 1606 1607 1608 1609 1610 ...
#>   .. .. .. ..$ : int [1:200] 1801 1802 1803 1804 1805 1806 1807 1808 1809 1810 ...
#>   .. .. .. ..$ : int [1:200] 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 ...
#>   .. .. .. ..$ : int [1:200] 2201 2202 2203 2204 2205 2206 2207 2208 2209 2210 ...
#>   .. .. .. ..@ ptype: int(0) 
#>   .. .. ..- attr(*, ".drop")= logi TRUE
#>   ..$ genesorteR:'data.frame':	24734 obs. of  5 variables:
#>   .. ..$ Gene.Name              : chr [1:24734] "PCAT19" "PPAP2C" "DEFB1" "TRPV6" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:24734] 0.00287 0.03161 0.07263 0.05734 0.00077 ...
#>   .. ..$ Adjusted.pvalue        : num [1:24734] 1.34e-05 1.34e-05 1.34e-05 1.34e-05 1.34e-05 ...
#>   .. ..$ Specificity.Score      : num [1:24734] 0.805 0.755 0.73 0.728 0.71 ...
#>   .. ..$ Cluster                : chr [1:24734] "11" "1" "1" "1" ...
#>  $ gene_snn_res.1.8:List of 2
#>   ..$ presto    : gropd_df [3,000 × 5] (S3: grouped_df/tbl_df/tbl/data.frame)
#>   .. ..$ Gene.Name              : chr [1:3000] "GCG" "TTR" "CHGB" "TM4SF4" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:3000] 4.43 3.16 2.15 2.1 1.88 ...
#>   .. ..$ Adjusted.pvalue        : num [1:3000] 1.88e-36 2.44e-48 8.86e-36 1.15e-38 2.76e-51 ...
#>   .. ..$ avgExpr                : num [1:3000] 7.27 6.57 4.85 3.74 3.39 ...
#>   .. ..$ Cluster                : chr [1:3000] "0" "0" "0" "0" ...
#>   .. ..- attr(*, "groups")= tibble [15 × 2] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ Cluster: chr [1:15] "0" "1" "10" "11" ...
#>   .. .. ..$ .rows  : list<int> [1:15] 
#>   .. .. .. ..$ : int [1:200] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. .. ..$ : int [1:200] 201 202 203 204 205 206 207 208 209 210 ...
#>   .. .. .. ..$ : int [1:200] 401 402 403 404 405 406 407 408 409 410 ...
#>   .. .. .. ..$ : int [1:200] 601 602 603 604 605 606 607 608 609 610 ...
#>   .. .. .. ..$ : int [1:200] 801 802 803 804 805 806 807 808 809 810 ...
#>   .. .. .. ..$ : int [1:200] 1001 1002 1003 1004 1005 1006 1007 1008 1009 1010 ...
#>   .. .. .. ..$ : int [1:200] 1201 1202 1203 1204 1205 1206 1207 1208 1209 1210 ...
#>   .. .. .. ..$ : int [1:200] 1401 1402 1403 1404 1405 1406 1407 1408 1409 1410 ...
#>   .. .. .. ..$ : int [1:200] 1601 1602 1603 1604 1605 1606 1607 1608 1609 1610 ...
#>   .. .. .. ..$ : int [1:200] 1801 1802 1803 1804 1805 1806 1807 1808 1809 1810 ...
#>   .. .. .. ..$ : int [1:200] 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 ...
#>   .. .. .. ..$ : int [1:200] 2201 2202 2203 2204 2205 2206 2207 2208 2209 2210 ...
#>   .. .. .. ..$ : int [1:200] 2401 2402 2403 2404 2405 2406 2407 2408 2409 2410 ...
#>   .. .. .. ..$ : int [1:200] 2601 2602 2603 2604 2605 2606 2607 2608 2609 2610 ...
#>   .. .. .. ..$ : int [1:200] 2801 2802 2803 2804 2805 2806 2807 2808 2809 2810 ...
#>   .. .. .. ..@ ptype: int(0) 
#>   .. .. ..- attr(*, ".drop")= logi TRUE
#>   ..$ genesorteR:'data.frame':	27458 obs. of  5 variables:
#>   .. ..$ Gene.Name              : chr [1:27458] "PCAT19" "MYCT1" "FLT1" "GSTA2" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:27458] 0.00309 0.01127 0.05123 0.05654 0.00399 ...
#>   .. ..$ Adjusted.pvalue        : num [1:27458] 1.37e-05 1.37e-05 1.37e-05 1.37e-05 1.37e-05 ...
#>   .. ..$ Specificity.Score      : num [1:27458] 0.805 0.751 0.726 0.71 0.704 ...
#>   .. ..$ Cluster                : chr [1:27458] "14" "14" "14" "7" ...
#>  $ gene_snn_res.2  :List of 2
#>   ..$ presto    : gropd_df [3,000 × 5] (S3: grouped_df/tbl_df/tbl/data.frame)
#>   .. ..$ Gene.Name              : chr [1:3000] "GCG" "TTR" "CHGB" "TM4SF4" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:3000] 4.44 3.14 2.12 2.08 1.88 ...
#>   .. ..$ Adjusted.pvalue        : num [1:3000] 1.09e-35 1.05e-45 1.15e-32 2.56e-36 4.04e-49 ...
#>   .. ..$ avgExpr                : num [1:3000] 7.32 6.58 4.83 3.74 3.4 ...
#>   .. ..$ Cluster                : chr [1:3000] "0" "0" "0" "0" ...
#>   .. ..- attr(*, "groups")= tibble [15 × 2] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ Cluster: chr [1:15] "0" "1" "10" "11" ...
#>   .. .. ..$ .rows  : list<int> [1:15] 
#>   .. .. .. ..$ : int [1:200] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. .. ..$ : int [1:200] 201 202 203 204 205 206 207 208 209 210 ...
#>   .. .. .. ..$ : int [1:200] 401 402 403 404 405 406 407 408 409 410 ...
#>   .. .. .. ..$ : int [1:200] 601 602 603 604 605 606 607 608 609 610 ...
#>   .. .. .. ..$ : int [1:200] 801 802 803 804 805 806 807 808 809 810 ...
#>   .. .. .. ..$ : int [1:200] 1001 1002 1003 1004 1005 1006 1007 1008 1009 1010 ...
#>   .. .. .. ..$ : int [1:200] 1201 1202 1203 1204 1205 1206 1207 1208 1209 1210 ...
#>   .. .. .. ..$ : int [1:200] 1401 1402 1403 1404 1405 1406 1407 1408 1409 1410 ...
#>   .. .. .. ..$ : int [1:200] 1601 1602 1603 1604 1605 1606 1607 1608 1609 1610 ...
#>   .. .. .. ..$ : int [1:200] 1801 1802 1803 1804 1805 1806 1807 1808 1809 1810 ...
#>   .. .. .. ..$ : int [1:200] 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 ...
#>   .. .. .. ..$ : int [1:200] 2201 2202 2203 2204 2205 2206 2207 2208 2209 2210 ...
#>   .. .. .. ..$ : int [1:200] 2401 2402 2403 2404 2405 2406 2407 2408 2409 2410 ...
#>   .. .. .. ..$ : int [1:200] 2601 2602 2603 2604 2605 2606 2607 2608 2609 2610 ...
#>   .. .. .. ..$ : int [1:200] 2801 2802 2803 2804 2805 2806 2807 2808 2809 2810 ...
#>   .. .. .. ..@ ptype: int(0) 
#>   .. .. ..- attr(*, ".drop")= logi TRUE
#>   ..$ genesorteR:'data.frame':	27481 obs. of  5 variables:
#>   .. ..$ Gene.Name              : chr [1:27481] "PCAT19" "MYCT1" "FLT1" "GSTA2" ...
#>   .. ..$ Average.Log.Fold.Change: num [1:27481] 0.00309 0.01127 0.05123 0.05654 0.00399 ...
#>   .. ..$ Adjusted.pvalue        : num [1:27481] 1.37e-05 1.37e-05 1.37e-05 1.37e-05 1.37e-05 ...
#>   .. ..$ Specificity.Score      : num [1:27481] 0.805 0.751 0.726 0.71 0.704 ...
#>   .. ..$ Cluster                : chr [1:27481] "14" "14" "14" "7" ...